pkg: finder -> resourcemonitor
This PR drops the existing pkg/finder and replaces with
pkt/resourcemonitor. The functionality of the package is preserved with
minimal API change, and no change of behaviour is intended for
`nfd-topology-updater`.

Rationale for this rewrite is that `pkg/finder` mostly grow up over
time with no organic design and as aggregation of feature.
Thus, the design was messy and the code was hard to follow and to test.

The current rewrite wants to preserve the responsabilities and
the basic design tenets of the package, but aims to provide
a much cleaner interface, better readability and improved testability.

Signed-off-by: Francesco Romani <fromani@redhat.com>