Pull Topology Manager Policy from /configz endpoint
The Topology Update will try to pull the TM policy from configz.
If this fails it will try to the read KubeletConfig file.
This may be more stable in clusters with mixed kubeconfig setups.
The configz endpoint is as unversioned API and it may change in future.
Configz also requires new RBAC rights to be given to the NFD container.